Reduce number of Helper methods
