+++ platform-tooling-support-tests/src/test/java/platform/tooling/support/tests/ArchUnitTests.java
@@ -26,9 +26,11 @@ import static com.tngtech.archunit.lang.conditions.ArchPredicates.are;
 import static com.tngtech.archunit.lang.conditions.ArchPredicates.have;
 import static com.tngtech.archunit.lang.syntax.ArchRuleDefinition.classes;
 import static com.tngtech.archunit.library.dependencies.SlicesRuleDefinition.slices;
+import static java.util.stream.Collectors.toSet;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static platform.tooling.support.Helper.loadJarFiles;
 
+import java.io.IOException;
+import java.io.UncheckedIOException;
 import java.lang.annotation.Annotation;
 import java.lang.annotation.Repeatable;
 import java.lang.annotation.Retention;
@@ -36,7 +38,8 @@ import java.lang.annotation.Target;
 import java.util.Arrays;
 import java.util.Set;
 import java.util.function.BiPredicate;
-import java.util.stream.Collectors;
+import java.util.jar.JarFile;
+import java.util.stream.Stream;
 
 import com.tngtech.archunit.base.DescribedPredicate;
 import com.tngtech.archunit.core.domain.JavaClass;
@@ -50,6 +53,9 @@ import com.tngtech.archunit.lang.ArchRule;
 import com.tngtech.archunit.library.GeneralCodingRules;
 
 import org.apiguardian.api.API;
+
+import platform.tooling.support.Helper;
+import platform.tooling.support.MavenRepo;
 
 @AnalyzeClasses(locations = ArchUnitTests.AllJars.class)
 class ArchUnitTests {
@@ -133,9 +139,22 @@ class ArchUnitTests {
 
 		@Override
 		public Set<Location> get(Class<?> testClass) {
-			return loadJarFiles().stream().map(Location::of).collect(Collectors.toSet());
+			return loadJarFiles().map(Location::of).collect(toSet());
 		}
 
+		private static Stream<JarFile> loadJarFiles() {
+			return Helper.loadModuleDirectoryNames().stream().map(AllJars::createJarFile);
+		}
+
+		private static JarFile createJarFile(String module) {
+			var path = MavenRepo.jar(module);
+			try {
+				return new JarFile(path.toFile());
+			}
+			catch (IOException e) {
+				throw new UncheckedIOException("Creating JarFile for '" + path + "' failed.", e);
+			}
+		}
 	}
 
 	private static class RepeatableAnnotationPredicate<T extends Annotation> extends DescribedPredicate<JavaClass> {
